Completed 02-holy-grail-mockup
Only used css grid. I learned that I can
use css grid within a css grid. This helped a lot
to place the elements where I needed them to be.
I also learned how good it is for a resposive website.
I learned what is the purpose of autofill and minmax.